feat: crlist working

+++ Codigo/src/is_prime/main.c
@@ -4,9 +4,9 @@
 
 int main(int argc, char **argv)
 {
-  int n = atoi(argv[1]);
+  long n = atol(argv[1]);
   int rest = 0;
-  for(int i=2;i<=n/2;i++)
+  for(long i=2;i<=n/2;i++)
 	{
 		if(n%i!=0)
 			continue;
@@ -14,8 +14,8 @@ int main(int argc, char **argv)
 			rest = 1;
 	}
   if(rest==0)
-		printf("%d is a prime number.\n",n);
+		printf("%ld is a prime number.\n",n);
 	else
-		printf("%d is not a prime number.\n",n);
+		printf("%ld is not a prime number.\n",n);
   return 0;
 }
